Title: 두 수의 거듭제곱, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240901/두 수의 거듭제곱/two-numbers-of-squares.cpp
@@ -4,8 +4,8 @@ using namespace std;
 
 int main() {
     // 여기에 코드를 작성해주세요.
-    int a,b;
+    long a,b;
     cin >>a >>b;
-    cout << pow(a,b);
+    cout <<(long long) pow(a,b);
     return 0;
 }
